renamed modules in _select-location/
