+++ src/routes/tools/local-climate-change-snapshot/_select-location/LocationMap.svelte
